changed : new makefiles for testing to support cunit and other methods
r3116

+++ FreeBASIC/tests/dim/submod/submod_b.bas
@@ -1,11 +1,11 @@
-
+' TEST_MODE : MULTI_MODULE_TEST
 
 
 #include once "submod_c.bi"
 
-	dim shared ext_statarray(1 to 21, 1 to 2)
+  dim shared ext_statarray(1 to 21, 1 to 2) as integer
 
-	redim ext_dynarray(1 to 21, 1 to 2)
+  redim ext_dynarray(1 to 21, 1 to 2) as integer
 
 public sub init_b( )
